requirements.txt
Django
gunicorn
whitenoise
sentence-transformers
numpy
google-generativeai
